Support field positioning offset and partial rendering
This patch adds support for custom ^-- command which can be used to
split the zpl stream into parts that are rendered conditionally, for
example based on value of some template parameter.

This patch also adds a way to specify a global coordinate offset when
rendering the label.

Also fix a bug where rendering multi-line text could skip the last line.